persistenceservice/src/main/java/com/eone/distributed/persistence/service/PasswordService.java
package com.eone.distributed.persistence.service;

import org.springframework.stereotype.Component;

import javax.crypto.SecretKeyFactory;
import javax.crypto.spec.PBEKeySpec;
import java.security.NoSuchAlgorithmException;
import java.security.SecureRandom;
import java.security.spec.InvalidKeySpecException;
import java.security.spec.KeySpec;
import java.util.Arrays;

@Component
public class PasswordService {

    public HashAndSalt hash(String password) {
        SecureRandom random = new SecureRandom();
        byte[] salt = new byte[16];
        random.nextBytes(salt);
        KeySpec spec = new PBEKeySpec(password.toCharArray(), salt, 65536, 128);
        try {
            byte[] hash = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1")
                    .generateSecret(spec).getEncoded();
            return new HashAndSalt(hash, salt);
            //byte[] encode = Base64.getDecoder().
            // System.out.println(encode);
        } catch (InvalidKeySpecException e) {
            e.printStackTrace();
        } catch (NoSuchAlgorithmException e) {
            e.printStackTrace();
        }
        return null;
    }

    public boolean validate(String password, HashAndSalt hashAndSalt) {
        KeySpec spec = new PBEKeySpec(password.toCharArray(), hashAndSalt.getSalt(), 65536, 128);
        try {
            byte[] createdHash =  SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1")
                    .generateSecret(spec).getEncoded();
            return Arrays.equals(createdHash, hashAndSalt.getHash());
        } catch (InvalidKeySpecException e) {
            e.printStackTrace();
        } catch (NoSuchAlgorithmException e) {
            e.printStackTrace();
        }
        return false;
    }

}
